got 3 Document functions to work: Copy, CopyData, and CopyView
Copy - copies the data and the layout of an object
CopyData - copies just the data part of an object
CopyView - copies just the layout part of an object.
These require that objects have an inheritance hierarchy that ends in an "abstract" prototype that all the layouts point to -- if the layouts point to.  Abstract prototypes are indicated by having a field AbsractInterface