Tyler - you might want to look at this.   Needed to fix to get umpire example to work
in the umpire example, switching game assignments causes the list of assigned games to continue to grow because DocumentViewModels in CollectionViewModel is never cleared and so the documents keep getting appended.  Changing to Update clears the old models, then adds everything.

It seems to me that Add or Remove should only be used if we know that's what happened -- otherwise, Update is safer?